pyproject.toml
[project]
name = "discord-music"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-ext-voice-recv>=0.5.2a179",
    "discord-py>=2.6.4",
    "fastapi>=0.128.0",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "python-vlc>=3.0.21203",
    "tqdm>=4.67.1",
    "uvicorn[standard]>=0.40.0",
    "yt-dlp>=2025.12.8",
]
